app/requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.1
Pillow==7.0.0
scipy==1.4.1
six==1.14.0
torch==1.4.0
torchvision==0.5.0
Werkzeug==0.16.0
